refactor(native): 命名統一 snake_case + props 化路徑 + 清樣板殘留
命名統一（snake_case 對齊其他 native 專案；class 名維持 PascalCase）：
- 模組整組改名 GetPICoaterBackground → get_picoater_background
  （資料夾 / vcxproj / .filters / RootNamespace / sln 顯示名）
- 源檔 Module_GetPICoaterBackground.{hpp,cu} → picoater_detector.{hpp,cu}；
  _CPU.cpp → src/picoater_detector_cpu.cpp（歸位 src/）
- benchmark PICoaterModuleTestsFast.cpp → picoater_module_tests_fast.cpp
- 亂碼註解（Big5 殘留）改寫成乾淨英文 → 消模組檔 C4819

路徑 props 化（消除 ../../）：
- repo 根 Directory.Build.props 新增 $(NativeRoot)=$(EngRoot)src\native\
- vcxproj 檔案參考 ..\..\ → $(NativeRoot)；#include "../../..." → "modules/..." / "pipeline/..."

清樣板殘留：
- picoater_api.vcxproj 刪 FeatureA/FeatureB/CoreLib/GeneratePcbSegMask\include
- benchmark / core_cv_benchmark vcxproj 刪 $(ProjcetDir)/$(AoiTestUtilsPath)/
  $(SdkCommonPath)/$(CommonLibPath) 殘留；修正錯誤的 Module_GetPICoaterBackground\include
  路徑（實際是 get_picoater_background）；修 $(ProjcetDir) typo

+++ src/native/modules/get_picoater_background/src/picoater_detector_cpu.cpp
@@ -1,9 +1,9 @@
-#include "Module_GetPICoaterBackground.hpp"
+#include "picoater_detector.hpp"
 #include "core_cv/imgproc/core_background.hpp"
 #include "core_cv/imgproc/core_utils.hpp"
-#include "cpp_utils/timer_utils.hpp" // 用於計時
+#include "cpp_utils/timer_utils.hpp"  // timing
 
-// 如果需要 Debug 存圖，請解開以下註解
+// Define DEBUG_SAVE_IMAGE to enable debug image dump.
 // #define DEBUG_SAVE_IMAGE
 #ifdef DEBUG_SAVE_IMAGE
 #include "stb/stb_image_write.h"
@@ -19,44 +19,39 @@ namespace picoater {
     ) {
         if (m_width == 0 || m_height == 0) return;
 
-        // 1. 確保 CPU 記憶體已分配
-        // 如果 Initialize 沒被呼叫或只分配了 GPU，這裡做 lazy initialization
-        if (h_col_mean.size() != m_width) {
+        // Lazy CPU buffer allocation (Initialize only set up the GPU buffers).
+        if (h_col_mean.size() != static_cast<size_t>(m_width)) {
             h_col_mean.resize(m_width);
         }
 
-        // 使用計時器來觀察 CPU 版效能
         {
             TIME_SCOPE_MS("Total Run Time (CPU)");
 
-            // 步驟 1: 計算列平均 (Column Means)
+            // Step 1: column means with outlier removal.
+            // stride = m_width: h_in is tightly packed by column (no padding).
             {
-                // TIME_SCOPE_MS("   1. Calc Column Means (CPU)");
-                // 這裡 stride 傳入 m_width，假設 h_in 是緊密排列的 (無 padding)
                 core::calcColumnMeans_RemoveOutliers_cpu(
                     h_in,
                     h_col_mean.data(),
                     m_width,
                     m_height,
-                    m_width, // stride
-                    bgSigmaFactor // sigma_threshold (雖然名字是 bgSigmaFactor，但在此上下文中常被共用)
+                    m_width,        // stride
+                    bgSigmaFactor   // sigma_threshold
                 );
             }
 
-            // 步驟 2: 計算背景與 Mura
+            // Step 2: column background + mura.
             {
-                // TIME_SCOPE_MS("   2. Calc Background & Mura (CPU)");
                 core::calcColumnBackground_u8_cpu(
                     h_in,
                     h_col_mean.data(),
                     h_mura_out,
                     m_width,
                     m_height,
-                    m_width // stride
+                    m_width         // stride
                 );
             }
         }
-
     }
 
 }
